Work on test coverage.

+++ src/main/java/de/tonypsilon/bmm/backend/team/service/TeamService.java
@@ -1,7 +1,6 @@
 package de.tonypsilon.bmm.backend.team.service;
 
 import de.tonypsilon.bmm.backend.club.service.ClubService;
-import de.tonypsilon.bmm.backend.exception.AlreadyExistsException;
 import de.tonypsilon.bmm.backend.exception.BadDataException;
 import de.tonypsilon.bmm.backend.exception.NotFoundException;
 import de.tonypsilon.bmm.backend.exception.SeasonStageException;
@@ -56,27 +55,44 @@ public class TeamService {
                 teamCreationData.number()));
     }
 
+    /**
+     * Only to be called from another service after deletion
+     * of the players of the team has been taken care of!
+     * @param teamId
+     */
     @Transactional
     public void deleteTeam(Long teamId) {
         Team team = teamRepository.findById(teamId).orElseThrow(
                 () -> new NotFoundException("Es gibt kein Team mit ID %d!".formatted(teamId))
         );
+        if (!seasonService.getStageOfSeason(team.getSeasonId()).equals(SeasonStage.REGISTRATION)) {
+            throw new SeasonStageException("Saison ist nicht in der Registrierungsphase!");
+        }
+        if(!team.getNumber().equals(getMaxTeamNumberForTeamsOfClub(team.getSeasonId(), team.getClubId()))) {
+            throw new BadDataException(
+                    "Es kann nur das Team mit der höchsten Nummer gelöscht werden!"
+            );
+        }
         teamRepository.delete(team);
     }
 
     private void verifyTeamNumber(TeamCreationData teamCreationData) {
-        List<Integer> teamNumbers = teamRepository.findBySeasonIdAndClubId(
-                teamCreationData.seasonId(), teamCreationData.clubId())
-                .stream()
-                .map(Team::getNumber)
-                .toList();
-        // TODO Log if team number sequence is not good.
-        Integer maxTeamNumber = teamNumbers.stream().max(Integer::compareTo).orElse(0);
+        Integer maxTeamNumber = getMaxTeamNumberForTeamsOfClub(teamCreationData.seasonId(), teamCreationData.clubId());
         if (teamCreationData.number().equals(maxTeamNumber+1)) {
             throw new BadDataException(
                     "Das neue Team hat nicht die passende Teamnummer. Erwartet: %d. Tatsächlich: %d."
                             .formatted(maxTeamNumber+1, teamCreationData.number()));
         }
+    }
+
+    private Integer getMaxTeamNumberForTeamsOfClub(Long seasonId, Long clubId) {
+        List<Integer> teamNumbers = teamRepository.findBySeasonIdAndClubId(
+                        seasonId, clubId)
+                .stream()
+                .map(Team::getNumber)
+                .toList();
+        // TODO Log if team number sequence is not good.
+        return teamNumbers.stream().max(Integer::compareTo).orElse(0);
     }
 
     private TeamData teamToTeamData(Team team) {
